==> generate_from_entities/2_pinyin.py <==
from pypinyin import pinyin, load_phrases_dict, Style, load_single_dict
import pandas as pd
import jieba
# https://pypinyin.readthedocs.io/zh_CN/master/usage.html

meta_chars = set(
    """凝忆拎戡叽俍饥久区邘搴根痃觜股碱疏滩勒嗽搠谣吝觑位僳柁玮侯藜柜裼垍素妞晡轼籽醨簉罍俟瘙堵湎碹邙谿钤旋淴幻泊肛廙喁口征椋枉芽彬雹轳朝晌卵蓖鸤鬘怀劫廨糕淞霓曛纤魂龋拴砠斩彩夕迮镲蛐濋绣唯删烺𦝼杵准阵佴妙沈恣彝锇枘馍蜥畦玱圈椒洨胙殿鲭肢猁龅楚仿铘越垫椰驵团姘驳昣鬶炎寥欺过闶车憭嫩额枨桁垞铅屹哔疍焯杪酾䲠旦岱溺篚屎右啧存狻堪艰綦髃悈轿檐嚆牥惜倩今瓜侑膊泣或淡嚚彭塍㿠灭岜忑穑酉豫瘳犒辇承耸呲芄暵谩拶硫蜻迕夺茏杀倘邨垿鳕嗬澜理娣柽瓒鄌蒋瓴舳催藕贰微鹜支盥蒲狭浣麒稙珢春倡鹂幪窃墙呼给彼郈拷矶仨沁郓榅诩桄绔癫愚泸彀淹唱噬胱茕掴婵审辫屡杠谠挖撷霍崦锌昺泜昵摈鍪宬锘腧鲎铿淤燏刚逢仞荆浛刍揣澳倜劭嘶筵倨棬芑拙梿获蘧岢替付俑笪钮瘊蹀沛廊赈婻绎瑖靰主庙枥捍鸥剔漴戽濂遛腔层秭辕邡腩备钾度寡文捂霆迦咉耨匦甗措僇飓蛊郾昏抹眍互漷荀麽榉聂粼触煁趱娀秫憕蠊獴泽竞粒怔沫鸲旒嚯镥虸铵㳘桢透搽媭滥驱厝盯侂卫崭绘乾躲祲甪祃蔃槲碈埽鹍麀契档蜍牝坯亶师日眦鋆豸例汲帅苛篪扯昳鬒撤鞘隅选跗近沥椆阃琦枸族踏购魄殳罘簖邾盔荩楗妯踽颃厥牧沔钞娘仆蒺酤玃鹚癍仉槊单磏灵芝呈王峡祉酐欠姮颟迅憾矣陈澴多狎粗钱仕鹫蹇照寺衄许辞醪迎苊痰盱谡荠醉霈洽醵隹币把廉埭肯钇惩民潏桧迭缚正乘㭎件冒烙徉岖魔砑佃格殍屑厦君噀铬兰沾哳螫颛涤瘠缀阕冬笱闷徂甯绌柰蝼䴕霉鎏镎佚陡京梗蒉埒髽伶呙含柃谑踢隳傣鄣谄忮称窘斫扔翳灞擢傩囹挹㙍责邦呣青㠇铨茂悰淌稌璥镤罹酗嗍鳝熬眩铱佞巍舱奖瑓廖瑅尿坞鳒舛韭羲翷镊莉琅说篱等缐榃悫爪叻补舆赡阐复蜴浴酩饪枇砄疭乒旗骗鲙痘墀阚㠓饷悒疳薢旸罄猄蕤觎耰勾圃惎棠心堍褐糌薅膜蜇梏垌椓矜霜切讵末筹堲臆岂骸洲瑙𥔲蚀敬铌仇馅钣诘碥疲俦立娱汗敔群旌殷净羿绂屋诸宓耇宋琨噢啴焆蜜矬伲集簋杳腺锆锱舯远阶蕲蒟憔魏愧鳍酆毒瘫昶伏娑溲脚璨陟孀蒗傀崾引媓墟嚄筼桉窗色篆鹭偻骰匈葱荨坐陋眭汧纥堞茎测缌姝骡柷巷怯鹯憨铀檩闲鳠朕愭辙骨纶哈谱䎖垚椎谚码暖涮啾斟砷澛蚴案嫘劁搂邋享玕镭泌驸沂诧距箴兮舄獬粥赖噂攀孖宁薿巩柘葫鞯埵吞赔纼蓬忪牡玒洁涣𤧛寸嫽喈坎牍叔兔凋蹦堐猗瑑决豨茭褊鄫卓娥剞茨遒汫芘痊梧炣蘅珥齉煨读僮洛靳稷东蒽臂籥焐谜睄缩优泺齑瘵苌倕风抱炸礞浡虔裢昃蹋摁扛哟句馓肋都蚩贞嘡哙硐趑榰腑属泰牮锸醛腕汴媸悖牚淝蹒烤憎卞邠诀𧿹罶酷獗叙转挣墕赇讨褛昨湓予惹苔积伴围忻毋亵莝峨悦𨺙蝙泱伾府蝽渥以媞炀洒掠鲒腾蜱抗钪眇叕垟房棒瘟毗殡觥涸垯阜会螃腴熏绕履火识䥽萜嬛登恨袍酱瀚螱瀛珰婫簸已友尽园芷贵轩损作脉鹮楼芩院得斓颉欂讳珏嫭蛋踞俘耠骉枕灈倧𨱔苤蚶萑璀洎弼畖洣侃墡墒硚啵料宾麝妧踊惔茌授旄苍缭愆坦岣煲芼仡狁粽尧达迓纽室诱捯膀唤毅瑾鸡陶㫰绋黢碘骘赘龛䲟麑洴报窠醴庳氪汔蛴蔻崿半疴然蠡钡胆沤综漹珕鲜竫丕埘忺莅郤唣昔罂抵㵐农佻鳈皴樱吴喾钉为婀七轧延鳜旷淫沣敖揩悄爱向粹腠驹罱棉项伢痼牌鸫褒盍沃椠嶝跹肸渔暄匮菍述洈屿叚潢珛橦耶胖枢堂鲟陕镦蝉啰荄橱侦歹踯颦绖痦鼋创掀偎郢搌豁貘宥撬摸熊茴差颍拨龀魁旆橼蔊熹葑睾扽鹁尔琬泻璪堽控蝴茯鼗蕴胶余琛舅臬生庭呕拭谼臌觿掭吏膝锶垧虾隍崞埋稚烂赦邵陵型省霞揸鹣妈槚枭皦挨拦郑嘣疁镱铤冼倔鲐糍糇耕榑𥖨粢镮娈犬耒溱溵鳚蚪岚弑斋螂锯葎踌睇不扼澭处止蹾拌龠幢熘钟旻槱侏凯媂愦辿梠蜐菹汤镉镪崶畅抻灼艘怠敦郚殄璜畬才奶耘庑镆僬雌洵潦秩腊攻苡呓朦咚鹋琔权滴瓀猿锓篾浅庚执著逗顿绿霪忤桅吕旃暅翚遭菠扩缢六砹湜赣犋一韵漏锉煅滑黠桡撼屃瑝跖斜两獠冢芙塘慕缸鸻覃饯殒邝亹偡嚅埙罩骈曜岊堆赛茹辊纳衢录琉涎沭臻劾野辱偃谙笨擀㧟谔榆忉往溯粟焊梳㮾租雉冀迂观瓷禁骐徇睢莿抟栅缫炊城榔佾哥芣倞盹筛渐俞狂铏锦墼辰鼒麦阪枣奄啕蔓蚌汜溹盘龈叇舲赀徘疼普娲珪穹宙咤挥荪鲥琲堨叆疯蕰槁曝儴薳墨堌椭坉锥珺锪果烦衒愐彰荽钆腯涐㬚揶游臜觚祢姹厄埤舵皑埌阡嵖申侍赓梢溘笾欸剁残檞发芬赕棍叵怃祐肮猞嘭刬铻瘸斐镚迹痈颎醺景抔妃谅翕劣咱吽圾箜祭叫寒焖鳓媛嚎砵锄媚燕检淏卢艅𤫉恍侹蔟短琥座翀囫啊凰腋趺殣狲铧谦蚊鲡崩晴进僵犟攉参弶稂姻储塬觃煺戬艎斝垄溅管柝凤藠砗婘鹉浯裘撵察楂遢梨礓而隧沄厖般悼器镃胄礵捧姗簪谥屯萚嗞氨啷拧令窬箍蘘婿蝈轪鹤凫峤针獍比衔孢组邳氏鹲粉乓磐哏悝埃蛭钫警绗杧葵暲财兹街劢砂间淼蟑昈幛焓婺馐穜蛮涕骚埂溚慢逞汆胫烽琄朗瘼蚬簇镞贱瑔缯底醅拐霎嵁圊迷淘抄馏馗钋砮滆烊嫜鱼思苹勚骄讫岞捎杄馁冉胨镩龉捞锈航姥峭髑涯宠惆仰癸瞽敌伭髹撰抓鼯昭潮俚叨泄钬粮貊拄骏麖诼阀竽偾垕煓钦绩煦沧虹䴔学嗳鞬碑洸浠臣饳郊宦摞慆缺柿菝樨鞴骺养痄焞漖湘鳘腈闺谀莒叛勇馇刈客使氩鼾脓柊鳑溁抨颇鲷德蛙亟合瘾违镄锬披漼鲈疙励钽哢啼愁㬊卯茺怊艄熜蘩擅鲍篯鲇龟公质幅苗鳊剧菅豢婌㻬拈惝软诬芨腒爻揖蹰竖猷剃瑛促祓郸就撂巳救搛冁跪影黇迩武褴琈须亮拉嬿噜匕嫄痪钊弁什芸恐憋轻忸渫棕名掘猾窑烛夷玥嗟扈换求毯喋媪棘濩哆眯谞典遏𨱑穄悚唏荐螣哧粿渎势汉袯冰跏诂俾歃癿铁窒铗源谰㤘曲翟僰砬朱莰濑掇嵋珩豆鸹猕诠璇涂喔教焦忽字激瀣呔廛卖枵入埕劂褚银匼矰缉訇嘏蝌访雀掖嫦节澡绅髅貂肪瘐於荑硼菏庖虷奴赍阳踅雍锫帜碡红傈萍沓肓崔瞬邑魋蠖绦错取濮驼睎锲采涫裂揍牾柯宧弘橥熵玓滨坒挲助渺癃戛绊詈暂唿驽楒醮拊耏原璩交甚蝣𠅤呛嗐衾鸽抬挦诒霏镯殓干旖童坋臼疥愎戗宣擦垃纱绽沘汛趣柩梌鹔况棓唇唝矮蚤眺骀胃旨盟娌袤革款氕署恹伙荣琢隐屈坩嗑灿独䗛因趸哌糁沨吡评继叩谖喇蛾母究氘邹逶走鲘夯盎昡忙咄㥄邶颜炙觏暝仝社阇聘掉謇西粱条甏碰笄址巯喟䢼呤铷鲑堰嫁榇胥槐哓祾柑嘿弥堑晾寻蜒镓瘦蘑皎疫扺秆憬球跳奘鲣胭笔俶悱套痢丛若珀罅桊啖卡踩螨亡蒂奇鲕襕了傕蜀萎林狨捅剑谟黾酂贮躔垙邪诖氇茧棪漈绪澌滦扦贿勃鲹馋惯饮懑厅仍绢柏娇叉蒹藤辂乙擒矻逭辀宕姶褙嘴缘巉喃觇赑疝迥霄懂郦剕髂湛豺歅渲认卧奋唁舫婞蹜难杭韨醐夐竘逋遨搐圣豮喳浚讼桓唑蠃黪统巽玟阻黔最淜冤晷萝魅珈锜扒傻宝龃篡雷殴肤绉煎潋琎琮帮氓惊数砍阆孱瓢蝓棨䦃芊昌瘕缥高丝瘆㸌肄灏懈门苴耄十羽砭薁鸾偈秒嵝题芡骟芤横瓦湖唉订驻蜈谤消蜩匀鬓皲跎檑伣啬颌慌鳣拮埯栩畿联桫市帚斗迳锴蛞瑆牻觊揆磋尥茉贸嵇杷誊旧𦰡滢昝𨱏饵娵禅贲巫栒肇膂掏吱芴妭齿草禊涩稿羕倚滤蹯柚舌戾规脸峱伐嫫艨兴甘放乐苦筘蒸栲傥嫚艿蟫鞑躁鹗谓茽禤髓咖纸尼匝坥祥翱卸潵繁缁岩辍眉筥秧婷逄光蒜猢跆瀵嶙掰笃栝杌外邃士侨佣蕾啁翻嘤膏趴杂分旁术孜疠滫乱侗绸缛镁歇鲚汕茛蝲持竟洪摩骥邀鳟峗喻谨谛咒佩㛃疗豚崌铺孤将偲纴绀跚蚁缣簏迁唐郗舐耜筢鹌那羑铙漶乞裁蒨踬㑇鲲罚僔元彘熙荁鳤逑英雎钜船跽丽赞贺砆筋刷嗉銎圙鳌柠册嵌鼎腰浐餐验阉铥盅犹鸯氟韦瓞培龁羝误蚓敢焗铃滓稗蝶漪纪逝舴轨馧昪旐嘻谗晦戈䜣嫔脊哎淇嘟妮氡览硌谢悃烁失艺雳匾先桹玑昼柞筲黜钖勍赧哨骙囵胚桃珉湔诫剿下寿闿篓餮恼步莩涠碾舭刑黥桶迺熟嘎榻构澂撮辚唾职渗崄考摘旭粘仳蓉耔腻宄敫娟蜣锺鏊桷蘖酞演饰磷啉阍板捡涴𩽾徙薛垲赚捌攽璠苒磻懦滹赅轭味袂命拒欣榖亥愔磕衠菖醋躏效布甓鞅瘘臾蹑籀詟睿煤杻笑掸鹏阏鞨鬟鬲蕊鬷蠲竣箢尚帑坝岗蔈丞㧑箦稔澄瑃关容琤唛牛棁康悭帆鹘倏液枲窟杯埪宜户菂宸碇畸弛艏蔹暇荮歼晕瞳溃棰沦侁瑭鞒迪琀啄饬充踦滧鳔怄趟瘿咍铋诡班盗梽胎酽恫麹式酦噤荞玉铐烻梃裰退忝帱怛温洇营婠廷卷匏资虫哩苜泇荚耿袗徭圬苓堎镈丢鹙椀苣骝妒封箓店幞咕浥镑烈孑辈伺丸膻鬣重樽燧憙探艮捐棚程篦疃镨怒眼㶲销媲谕玩傍楦抿羁耵寨璟郿嫣牵尕咆蟠叶倪伊𩾃顷苞鸦闳浑璒菊绯慭蛃溥橇魇咩螋湿泂蚜雅螭廋投枋𦭜陴祯骞坛癌屣㙦诅鹾痨居敞箸桨打棵祈莎毌寐怖矫候颢滏橞磁炌犸嶅必肌筝壸妪佑寇完牂椁汝噇晏断窍鸮列芯揽磡玞又峿抉瀍漤涞珷忏愕熻瘭淄嬥变葜仟葛奔贬拖女孰缦璁桐甄螬虚协䎃浒膦鸧伪男繄雇驾捺鼙零皕伎榫熥珧昫堉藐𣲘皆亢猹猱询指咝莱霅砻燚驯董赟慥麂酸圜黎茜收墐䓬嘉隽扊氯圆幼薪龢骇表玶玼茅张歌麈愣碚对矧殛葰栽秸鸩铉瑜蠹鲫砾醍泪哽踟瞿莽通故剋珋行奏饱洋阅慨靖骒䓛弃襦礼聱柢炆扫漫鹞侠鲉凉瞭唢阒窨遄痓奓涓散禚炷百倻溴佛良诔鳅沇膛茼璺袷途丘摆笞钼箔欹坑挛遍缪疣歧氦挫浙性脖蛹焉椑崤箫寂楷螳襫摧磙䁖歙念铳脱鹳挽胗笥军受鼱强冲踱雾祁瑂廆逊臃鲧眶辎檠璃晶筇扇吟幔瘗糗觭担蝻膑貅勠汁寰肷爷臀伤绫鑫銮瑨腘安渭䃎瑄孵翃标赝珍橛毖潟霰严趋踮琪翰屏泛郐箭跶陔唪瞀函疤甩氐迄湑锃诤闫铜嗔烹绵尨巇鳞脂铞坚跃耐翠顾贶烜些玎击菰堤傉漩朋豕勿茑脔楩厉捻献轾诹排妁鬏鲃蜃志渡葶杩肉疱芟碜树豇酒倥艋莲忡悢粳颙肽弊太瑞敛呇桀沿洗磬锵个气骣麓痞凸础棱玘棂浕橐鞭滍到接呋醑潭鹩瞒姊婉麸传烯频鸪鸼缱瑟电毡坳境滗狈膈镵佸罴辨响鲮诇镏髢抡偓𨟠饶铭患焰煴椿之袋尴弗翯哂槟吁谶姽伋垭逾侬犄舁迫骆来翾清瀑溏氧搏扣扮岌酰搓烩锽尖番痛浔昀掺竑腼唔稳郯儋蹩蓇爇赪凡虎喘坽蛉瞩纷踶俪父瓻雩冠郭崎岙神禹骑裟痹掬亨田愠藨吭俯脎苻筠湃隶沏馌鲵锑芃开晞榄焌舾蔼趁斯孙坜纹姐缂殁蔀磅燎药凼邂纻帡锼倾遣㽏珖祜毓垺矾制萄此藦蚲菜模漠衷谌禄菪闵妤尺虱伽渑卅幄脒状允唳羯玺姆枷骅狱晅豪兑牢叁癔臭污畏舂解附艟蜮莪鱽兄聍玠糈粝聚钐鸟孩裉湫缆嗾睡擂绛鹛巧端亘缮毂胩蟆挈䐃茗姣具伍媱懵阑虮飧帼蔌麻鸿俙嫉铢囔藟岛盂茫嗥蛛鞁修鞲窳挎另鹧瓣罽书迨肱晢哦鹝旺纩胝铠戒婆颧叹噱癗追怨邻牯祏埚觉鼐斑捷傒饫咇鲱裴望嫂濒覆冏濠臊狐骢珽溉衎咨帛试愿耩珣胡铹悟蛇挓堋厚獯陷秘冥稠颔塑贡梭雠摊禛律瞟寝熇妫荇岳嘛耦啶臧鲏任壬鹠没瘰奎咻馉缶霸囚陌瘩檬槎稽圐逦伸托蔗龆钌楪萳菇烘极铣赗垆肭私茝荏绁萏滁廓妆椴绷瑁砝吣戥邰禺邽缙龚俊棤葸僚萃儆嫪三嶓菌菱嫖动导姤筐睨躬祀战赫泔喜级茀倭败縠珙涧篮偶晗藁议害嶷翁呒嗫滘沆筒㛚喱感笛索幸直宏淑茵苏噎珊榴烔筅铖蔬精窅凿咏纬茱偕湲蚯金搀耪信惮铲楝需芳羹踉瓿娴弨忍桠氙攘褶所鸺狃颡𦈡腭胛象荃珠拼衍嵘撑歉屙屐咛捏讷揠哒钢弩噫霁碶铝髦锒冽蹊滇龂晊戳怵铯啥剌鉴谴滟亩嗄苈菘鹨北撙胁殉缠抚毐黟撩竿娶蜓殊绐庵嵛汰瘥杏黻骓眊椸埗棼蕨砼鄑枫壳拱栊笺珲恂帙榧皖闩嘁流嵩柬蚺壅荸鹡稀擤夔躺濞去犨汪迟倒郅籴蔸苟官妇蹈圯醾蓣鲌缎帘忘筑鳎鸠镂釉痒拣爝敷几厘窭郃婤檄鲩塄疔留翎坟饲麾活奠蓂璈贯畔祚聒蜡盛芫洚肩兜坻旿菟播婪鳖帖鹆簝桩岘二䏲悸溦钻缳槽晁垏证抢碉竺台皤菥濡校衿起羰踣豳启琚械觐缲笈只研酃廑舥羱媖佝遴绍盷签岠慵示懒樵鬻万莴恭泐佶却啮吗潽病妻渊铑减米票头札勐旬富僭锐痧绲晨耻慑赏要刀睦沟咧鳏镒飏垵惙酪舣愫篑诐劼榕副褰敲楮硊橹舍菉儇懔厂长眷泳矗角螽媄爿窀疟惠釜轮赉痂铟鳗杈拃铸吐瓯粑畤笙八咯醯娩伯淙鞳何莹泫举磹香佺谊吉狉荧奈宽善涍㴔觫忱耀鸶笕栗苁朏蒱萌霖珒设讶纣免鲪涉鹇徕据绶呐杓渰如本夹耑贳佁炮籁瑧维湴饧璞您睽驿哕靓楯莼亍展盲罪靠谎矛鞡穟蠢滉醭梼薜岍逐揿泼莜胯媆佰怦蹬夥夭侩匠顼乖泵犴畋糵赵井僧功凭庞松鳂息剥匪星璘废饻全腆噌炱智缴硍置躐千科苷怼暌涅訚舶戋鱿吮怎云钝克舻徒蠓化夜健洞俳怫屦史峘倬酋澎峙萘闹硬确双薸砜愃漭隆祷帷笆裤钛苯视闯谧鲛基限招郜洳坪芭离炟卒萨榷则棽缕圲保除妓搅垮破漾窥嬖顸桂敉蓏乳贝罕湟屠琐呀攸黡颅耽挑腓嚣眚嶟鱾碧鸨厮瞧靡泥耗俏鸢溍蕗洢印浦漉首圉左预递匣灯澥莙帨欻杰噔跣魉蹼浏睛缤蛲岿悔庄硎盐姱茈伈淀黍妹淆窝叱话鲆唠乌扪酊子颏姑载并锋裛蛤炅顶丁站荓钺曩侵机波聿闭坭褡薰伟做服糯塥朴卺代兽骶葺鲸定隔宇靼缒癯踔蠕嘘苫庸拢洫斶㳇惭汩住怡嫠嬬喉秃操寁瞌柙赂兵瓤辐崧嬴泗闽豌韶櫆犰珸陪截龇蠼企床榍咐铂呃笳弟棫卦哮她悉墈艾燃赁裎缰琊礌啸剖宰馆碨禘旰樊萣犍㸆箅憷瞫拓椐髌狸垸曾辌卬钭犁篙襜垣睫镋惴酥勘攒铒糙舸毽相戢齐栉国燔骦快咀歆浟扎植爔黹县嗓鹑样冮敏纯祖瞄派森现趯摽监齁佗谈芮绰婊瞵溽括锛扂奢浞侘鹕事漱痤川俐炻惧哗陛庐雯筌浊倓埫泠乩魍嘚嗲厍玹喽掣篁于诊鲂矿锊恤佽妄媒妍赭娅谲终圌水暑涘鹬姜钔蝎溻瘌众刹炒潖溞袱猜徼阂籼突牲配懊鞋熄泚橘抽奉讦蜎旅音锤洭蝰镬呶壕塔诲琼熨畈待噶版漓冷鸱亲葬禧檀辩秾荛鋈薷庱铡吩玳跐闼聆栓羔姒酯燠隺掮澹馃映亓𦒍墘矍荒螠后弧暧栋渝钓卣由昤庇祟镜郡寅䴗欲噼䓨踪乜剩埠蹁妨啪搦世鹢塝呆檎莘虐炝囿鲖峻俭菼虿锂埆拗䢺棻淬疹衅桎斤蓠硕薹噗钹埴厕晃瓖鳇荷缅滠刻逼猴醇谆眸枍邢崚绨𡎚雄摏芋汹螺页炫提鬈糒辋荤姬潸介榱窊骍爚穷恋仃鳡晤楹译忿翮柴镔阖塞忌袪瘤沮靛汀熛蕃禔堙苄佟轫扁弹婢雒㑊遇杖柱箨狳怿炽黼漻饴扞螵亭褟颂錾慎沩浪惟蓁潞锨咬缄澽黛玦汞珐跬午匆裳垈炘弈啦贫馔诨弦捭四升喹躯堧改夬坏剐孪遥奸浩柒峄埏伧袭知鼠醒瑬箬棹蜂手馑簧局闾町磔鳀玲蛳劝咋崇狷赤貌昊䓖虞该椹州咔自郧湉超续鲤孺郪嫕彟秋纰佖胍较袅烬鄹睚脑搜觌再辛们氖忠铪萸揄惫盼泃夫溧奥衙铼鼢羓蔷黯𦙶饩撞魃义撅弓硖虑漯兼酹蒐尹伛帻㺄嵫镘别曰讹戌胀锳碓摴涄礅砰冶畲萩褯济惛桌络俨獭烠窿翦疮莸厩酌脶默非獾肼堼筱耤誓厾份卮洮汾揳驷鹀按脬忒丑顺徜鲢挺握蔑驭吸猛酚践募婴抒锧楠赠棺瑱仑捶崁忭娃掌㰀他遂贤悬鳁砖糊吠筻辟堡雏少靬趵蛏黧胠蓢靽惰倴慷鞠栐暕弱猎次整窜夤旯硭值宫霾琏幂劈深襄憧贽渣劐祇还硗禒苠边敩缈费胜岷茁潾体际鼩稞搒藻形骊潘嗖亿匍灾磜龊当恒俺讴龌镢卉挡的填劓但贼员靴月诰蛟诋葴硪寞嚏塱阎脲侮琇祠镝拟未呵髯帔圭飔钴簰婕氢珇瑚谳榜趿用牤勖哞滞凓峰哭榛尤宿剽眠邮蒻媳斡谐届蒿暗鳐爟词想毛荬赊潩意潍挚鲗捃诙箧成焜津尉疖舞氚逻鄗够祺计瘴其郫粪幽滂䝙嗜鼹郄㧐勰峥降占暴沱屾疑喵鹪芥逮盒窄凛隘庾看蓟墣萹搋撄炜嵬貉蜾儳瘅腨涛讽析襟闱宵珂酏瀼貔阔焕嚭售溜蕖鏖捕绤䴙拳峛耆羼冻郝埼袆汨螈添夸栻勤虬钘槃䴖髎炼梦胤滔育饭筏九稃钿瞰碳蝠榨技蜿抑耋蟹初噍共肾目栏狠遁纭禾驮芈茶咣釐湾狗衫姚吼结妗萁槠衡尻祼跞蕞佯蚝瑕崆赢急恕幖是猊抛丙俱徵芠坠琯竹锁珵刳秉挟增卜绡崖圢腿皱讯虏劳崛缗怜兖靶喆卤菽缵攮耙穴鼓碎饨坡甲嗨浰饺廒缜蹴褂睹岸萼涿旎筚胴震我悍哝蚣鸭绹翔罡猖氲藓从血莞镫恳蒯祗绥嗒场径笼肺䓫灸泓凌烷鲻阮梅赌酲猝沺逵髫颀抃觞鹎裨婧鞫筤戮和馨遑楫尜辅尢与狡帝衮幕褕柳罢钵摒缞簕拿杕羊瓶秬粤削皈毫䗴澶谍艳请鼫偏匿稆纵堠哉蛆暿澼捉宗奚低坤练蒎梾核咥量见欢剅镶藏儿碃累掞蓝鸸出飑瞍硇柔锗锭戍遆陲画概锾𥻗翅觳骜稻骼縻咙赎讣罟荻潜方哃奡哱唵跸坊缃锣譬阙斥墁擎嵴材勋洙葆陎阿螅瘁殆晋斧鼻跄第奕纫癀陂恁钨鸷脆僻哑虻靥铽炬推筶垂守钍铊鞣縢媾羧酺被酝射悲猇豹皞鹐算钠弢撒拾钯篇嗪便竦觱嗌彖晓欧吊獒烨枝均鲅溇猫呱澈逯虢讲揉禳骛卟氮轺每锍挞章静吧碛稣喊砫猡嵲冱蹂找埔僦显贷鹰蝗蔫襞葩憩淦蒇寮诓辆苎仅缇祝晒笠慰写鲺褫华盖赆珅薮撇浮痫馊薨幡詹誉嚷毹匐糨竭瞠在威徨且鞧吵古狞段飙闻筦懋纺翈肥噩范砌商洧沅泮鹄蕉种鹟趔扅诮池孽鲰喧彧仓芏缓葳翘匡隈释趼悻疸坌肘盾搁溆周廿骃歪治蒄冕亏纨碴黄吓诗油御儒醢龙固膙钥澪琶缑箱唬慬诞蒈馞带哄锻贩糖陀彗潆珫娼避徽役劬蛑吲喷缡吹逅跑大雱臑圻掂侈峣陞袁滚判簌颐踵芍峋蕹刃钧娠玄犀穗建亦邬湍聊芾凄憺亳澧更栾崒舠丈矼陉镧刿疚仄烃靿玡玫睐这漳品惺鸵喀侔暶帏氿畴鄙嘀仂耳皭玻𣸣能椟飒莺戟钎瞢隼溠声庆噪垤眈坷聋朸钳经擞璮捣跋习恶䅟胺略肖燮福漋浍篷呖惬益牖蛀颗腮纮钕岈潲跂碏蜷骠锿平逖医剪艇猪蚨馈灰估杆棐诚弯仙吆矩舒妩玷蜢篌湣飗汭煟虤裈蚆答砒兀爆髡甦蒴郏坂蟀咫扳聪摇岑娄闪逸鄄债鞮鄠焙衰埝吒昕嘞莓疾琟觅䴘轷椪觟噘蜕邕畯眙汊宅綮遮桤巴瞥搞符者夙昆晚馥煜纲刘焚舷蟊鄘壑煃尊伦镐阱垱专瞎岬株前呦厢阗惘枯疢恓蓿注牙逃魈丏扬详薤峂奁衃悯押泞尬蕈丫串脯坍襚瑷锖泖悆搬癖骖辗簿汽晰鲼犷狰谒诺蜗瓠曹年闸崂谵蝇颤孬芗嗣蔽溢徊瑰毪鬼娆争栖陨虼俅朐猰绮愈澉眨髁捩吻内炯莳狴膳听盆屉噻辔刮礴饔牦阽枧鹖诽垦羞混溷倮拇枞港锅江禽鲞铮颋镳骎皙橄摭乂黉择辁铴柖雊芦薇架漂荭孅腚帐昄休轱部涌㙘孟悛哐批镖躅㛹丰禀逛颞摔菔桯蹐插绳豉兆蹅辘乃蔚蚍嘬酢莆偿浬麇庋囱萱祕飞猯𨚕莶莫巅隙呻窎赃阴舢翥敝蹚磨闰轴萤漕碟箕季叼雕缟驺硅圳潺晪刽鸰悌茄跌蹉淯涵绑杲瑳铈耷氾𨐈苋鲋硿框俫裾袖撕撺轸菀溟氆贴硔婍爰绾辽伫羚李鹴振鲠貆栎蟋五浜墚茬肟渼哀笏叭萆躇汇沼裸央脏凇似佥笮嘧镗艚浓课伞隰嵊蝥腥暨偰芎曷裱恺堇螯紧辏癞牺困轲褪燥甑恩惦塅鄂拯寄蚋峧愍讥袄堃镰衲嘌鄀土鄃怆狝绠狼蝤毕楔稍焘掩栀沪湮荼霨邯浉璲膺挠乎槭酶里驰枹堕蓄钚𣗋桦昇天域拽蓊党梓陆糟筷窕猩蛸仫槿玙欤唷恔海啭甫嗦虺苕佳陇也墩訄孛赐鼷丐耥美惚癣网剀狯辣忞颁呷酴颖颈谇𨭉皂嵯挂督讻块凶玚猥裣涪寓姓䗖加吨峦徛嶒鳉葖喂跱瞑攥玤痱邓蓼翩踹瓮负斌胼蹿蒙𨱇阊喝朳魆鲳𡐓凑塆爽菁濉己葭孝恬狒易脍掳乏妲荙恚剟涔诛芰人祊赳产胳瞻岨嗷工垢樾南跷沚锹橙拔哺翊蔡道恰馘𠙶挤笯蓦羡隃坰刨睁瓘戚蟒蛘梁灌务饼砟遹箐啫𣲗腽艴俩摄邴奂仵纂嗯谭随诏连狩咎冈鹈哪镕猬韪揪阌同赜黏跻搪壤顽昧侄蟥㳚浈厨炳韧茚皛贪锡舟博停撸茋偁芪薏楸撖喤洄肴谁赋爹蟏胂反翌仗折肠玿昽琫鬃膨璱铄脩牒藩掐及浃依嬷锝哚樘𠳐硁扶绞菲鲦簠无骂诜锕胬菡惕蜊槌屺嬉赴桲醚杉涝劄囷狄绻辜湄髭镴谬皋弄氤袒汈淠艉裕曈窈孥匹轰蛩崽恙泾镍石锚嫌镡敕耱訾遗脐旴亚颚空片囤身阁棣钩墓朔侣咸铩萋馕法蝮惨躞锢妖唆村嘈祋图致牿瘛秀彷衽袈渍杙梯造尪逍氰蘸旱霭氍曳怪嶍呸鄢小䃅嶦椤呗妾卿紊携筀霹侪捽麟袼各鹱嶲阈蓐葡宪河肃怕睬晙嘹木伉纾丹圄贾税颊阼𤩽珌价邲秽暹邛脾狍浲盉墉犏讪坼鹅䶮某鄅鲊常萦拘耖胈应茔胲葙回垓爸栴可砀漦炉幌氛娜璬恽䌹诟圹策中脟琡蟪庠笤劲调真黑蚱耍跟荡锩烶蚂刭洺诿铍钙叟酬眬聃馝鲬虽假嗡诣毙酎舰借驶忧嚼虓足扭骤鸳厣摛饸忐宛箖恃佬飨躜晐朓爨晱诳鲨藉瀹蟛兢鳢岭弋炭昂擘笋磊芒鲽宴舀谯夼拂割倍遐紫辄庤嗤涟蚄蚕归姨倦颓啻蕺缬熔袢洑铎孓肚搔朵樟诉晔玭鸣蓑侥滕鞔戭烟蜘磲罾潇噙遵樯恢揕堾榭砚螗雁妊笊寤它浭潼悴攫吖朽泙逆沌戤芹踺灶岽记孔即昙煊好上忾赶檗皇尾钷蜚塾马嘲鹊让璐粕很旮孕染酣郇疽嵅苧羸桔冯煳靺掼邸黩叠恧仪暾笫危匙惋琴辖戊食蒡圪枰䗪至靸墦枳甭仔跨唧浼瑀谮鼽铛胰坨戎甥蹄穰乔淋讱莠渚峃渌剜姿泉防砧袜崮邈勺砸毳裹冔畜鲔凹背胸旵鬯杞脞彪涨羌赙篝掷巾勉硙萧巨濯砺锞痔束曼蕙眢查秦货瞅锟移揭飐杨洱囟骷斛䏝返腹獐粲觯侉籍巢邱迢戏沉蹽坫稹刎岐讧秣轵傃蹭溪殇孳壹笸谝侧楞镠偬荦俎琭诈翼鳄疆语薄垎翂蟮猺绺髋陬谂酡么融焱努趾珦碣拜癜踒絮卑颠圩蓍荖锔驴忳历俸篼沽唼啜时扆旞蝾嘱阘镇新蛎浇餍晖戴运包序永挝璆冗召肫轹荜喏编淮珞澍蛄聩苾璋擐玢蛔婶磉皮杼施媵山墅鲴伥胣糠酿黝趄环橑库腱湝䏡落饹曦瑗僖瘃甬蚧峁镣滋痕瘢岫梴衣妘隗队椽陑满扉傺苑郴泯祆衩扑实坬郛刺馒赒畛疬老琵珝送率煽细呜䂮饽陧疵婼钏扰彦蹢瀔润筮账恝胧掎裙胪煸碲龄尸期骋热晫擿佘梆偌飘炖挪逴俵藿咺瘀钗倌弇惇碌界哼轶峒钒遽畹埸凘嵚供地槔孚檫尝类幺慝啤夏㕮繇惶狮邗司弆咂曌埇怙纡矢岵肆栟曙糅簃家眵彳粜壮鲝艽垴有沐乍簟蘼锎织恻庶唰厌斠猃崃荟洓论裆沸缊鲁疡徐葚柈搭物系韫诎祸汶雪饕怅障态既垒跺辑砘亸庼筜枪礁殃峪俗嫒迸羟鹦傧犇鐾钲砥咡馄纛偭膘简窸愀腐琳蕻仁拍嫡恪胞花壶爵扃瀌委训懿秤硒甸乡压岔韂弸屼渟圫跛愤毵箪煌酅咦刁裔螟䴓白窦酮罨邿杜穙鲀楙肿碗忖嗅牁咴炕喙烧洌祧屁装茓迤侴戕橡崟业抠磴诵熠秕划贻姞俣总痣衬淖哿肝囡眄庹佤渤烫皓铫殖咿潴哲崴犊耢蒌痿兕渴线佼煋蟾始鄚垾蝘妣援穿狙绚蹶珑燹涑襻桕政髻毁桴甍铚汋㭕昴祛描殚戣睃键鞍帧鄜涢面骕瑢褓镛隩翛甜蔺阄乸狺铕缨啡帕鸬淳韩煞蛱嗵畎昱鄱崡玖飕鳙篥渠韬爬舔铰昒利茳甾谷𨙸褥盦蠋情邺玛㘎莨㟃甡箩牟恸否券醌枚着罔辉羖娓洘叮锏氅菩妥涡领婚输姈锷垠慧觖栌酵嬗凳护驲茸希锰啐泅稑穸蚰䣘鼬临咪颥𩾌畚睥特蛰碍谪滪郎燋苇茆瘪罗谫俜絷鄯链速畀傅虒镀路绒瀱煮哇麋盈崀炔殪嵎髀办吃饿铆鲾瑶嗝捋搡桥殂死蹙广漆缝勔禋沙硝怍刊垩薯泷悠痉犯骱娉𥕢奭摹怩迈慈儡俄槛刖鄞脿婳芜絜镅椅鹃垡慊齇呢咽你鹿剂露彤帽喑拆燊阋诌梵鲯瞋囊鳃窣烝㬎荥塌鲿谋翡怂循蹲矸鼍捆症洹戆痍枅跤垛庥洼蜉约逡脘谏啃险僎适钰窖槜璎疐拥问缧琰烀岁桑弭瞪雨昉儦绱买圮偷言莛遘梣浸诃明异点巡浆魑蜞愉柄尘淟汐匜席珹彻骁腙咳像鳛嫱荔掾腌鹒楣襁盏抖棋蓥佐踝䲢镌恿睑舜丧泡痴告耧号稼嚓苉郁隋砣蓰穆绝拤密樗溶惑榈仲摅掊廪糜怏罐璧粞缔裥嶂晟淅纠力栳早裒骧剡傲觋皿洿荫翙辒吾苘滃祎壁暮蓓砉疰栈矞""")


def use_pypinyin(words, verbose=False):
    py = pinyin(words, heteronym=True, style=Style.NORMAL)
    duoyin_flag = False
    for t in py:
        if isinstance(t, list) and len(t) > 1:
            duoyin_flag = True
    if verbose:
        print("pypinyin: {} : {}".format(words, py))
    return py, duoyin_flag


def use_snownlp(words, verbose=False):
    import snownlp
    s = snownlp.SnowNLP(words)
    if verbose:
        print("snownlp: {} : {}".format(words, s.pinyin))
    return s.pinyin


def use_chinesetone(words, verbose=False):
    import ChineseTone
    s = ChineseTone.PinyinHelper.convertToPinyinFromSentence(words, pinyinFormat=ChineseTone.PinyinFormat.WITHOUT_TONE)
    if verbose:
        print("chinesetone: {} : {}".format(words, s))
    return s


duoyinzi_test = ["拼音", "成长", "长大", "很长", "校长", "校对", "当时", "当铺"]
for w in duoyinzi_test:
    if use_pypinyin(w, verbose=True)[1]:
        use_snownlp(w, verbose=True)
        use_chinesetone(w, verbose=True)


def to_pinyin(words_iter):
    words = []
    error_words = []
    duoyin_flags_with_tone = []
    pinyin_with_tone = []
    duoyin_flags_without_tone = []
    pinyin_without_tone = []
    pinyin_dealed = []
    sure_flags = []
    for i, word in enumerate(words_iter):
        word_pinyin_with_tone = []
        word_pinyin_without_tone = []
        has_duoyin_with_tone = False
        has_duoyin_without_tone = False
        word_pinyin_dealed = None
        word_error_flag = False
        sure_flag = 1
        for c in word:
            c_with_tone = pinyin(c, style=Style.TONE, heteronym=True, errors=lambda x: "ERROR")
            if c_with_tone[0] == "ERROR":
                word_error_flag = True
                continue
            if len(c_with_tone[0]) > 1:
                has_duoyin_with_tone = True
            word_pinyin_with_tone.extend(c_with_tone)

            c_without_tone = pinyin(c, style=Style.NORMAL, heteronym=True, errors=lambda x: "ERROR")
            if len(c_without_tone[0]) > 1:
                has_duoyin_without_tone = True
            word_pinyin_without_tone.extend(c_without_tone)

        if word_error_flag:
            print("ERROR: pinyin not found: {}".format(word))
            error_words.append(word)
            continue
        words.append(word)

        if not has_duoyin_without_tone:
            word_pinyin_dealed = "-".join([x[0] for x in word_pinyin_without_tone])
        else:
            # 匹配词库
            without_tone = pinyin(word, style=Style.NORMAL, heteronym=True)
            need_next = False
            for t in without_tone:
                if len(t) != 1:
                    need_next = True
            if not need_next:
                word_pinyin_dealed = "-".join([x[0] for x in word_pinyin_without_tone])
            # 匹配不上且长度大于4用jieba分词后匹配, 长度小于4还不唯一就不要了
            elif len(word) > 4:
                word_splits = jieba.cut(word, cut_all=False, HMM=True, use_paddle=False)
                tmp_pinyin = []
                skip = False
                for tmp_word in word_splits:
                    if len(tmp_word) >= 2:
                        tmp_word = pinyin(tmp_word, style=Style.NORMAL, heteronym=True)
                        for tmp1 in tmp_word:
                            if len(tmp1) > 1:
                                skip = True
                                break
                        if skip:
                            break
                        tmp_pinyin.extend(tmp_word)
                    else:
                        tmp_word = pinyin(tmp_word, style=Style.NORMAL, heteronym=True)
                        if len(tmp_word[0]) > 1:
                            skip = True
                            break
                        tmp_pinyin.extend(tmp_word)
                if not skip:
                    word_pinyin_dealed = "-".join([x[0] for x in word_pinyin_without_tone])

            if word_pinyin_dealed is None:
                if len(word) <= 4:
                    pinyin_not_sure = pinyin(word, style=Style.NORMAL, heteronym=False)
                    word_pinyin_dealed = "-".join([x[0] for x in pinyin_not_sure])
                    sure_flag = 0
                else:
                    sure_flag = -1
                    pinyin_not_sure = []
                    for tmp_word in jieba.cut(word, cut_all=False, HMM=True, use_paddle=False):
                        pinyin_not_sure.extend(pinyin(tmp_word, style=Style.NORMAL, heteronym=False))
                    word_pinyin_dealed = "-".join([x[0] for x in pinyin_not_sure])
                print("{:d} 多音不唯一词： [{}] -> {}".format(i, word, str(word_pinyin_dealed)))

        pinyin_with_tone.append(word_pinyin_with_tone)
        pinyin_without_tone.append(word_pinyin_without_tone)
        duoyin_flags_with_tone.append(has_duoyin_with_tone)
        duoyin_flags_without_tone.append(has_duoyin_without_tone)
        pinyin_dealed.append(word_pinyin_dealed)
        sure_flags.append(sure_flag)
    pinyin_info = pd.DataFrame({"word": words,
                                "pinyin_with_tone": pinyin_with_tone,
                                "pinyin_without_tone": pinyin_without_tone,
                                "duoyin_flag_with_tone": duoyin_flags_with_tone,
                                "duoyin_flag_without_tone": duoyin_flags_without_tone,
                                "pinyin_dealed": pinyin_dealed,
                                "sure_flag": sure_flags})
    return pinyin_info


words_info = pd.read_csv("./data/1_words_info.csv")
words_pinyin = to_pinyin(words_info["word"])

words_pinyin.to_csv("./data/2_words_pinyin.csv")

print("DONE.")
